change EdgeDetectionOperator names
