+++ src/OpenCvSharp/Modules/ximgproc/Enum/GradientOperator.cs
@@ -8,17 +8,17 @@ public enum GradientOperator
     /// <summary>
     /// Prewitt operator.
     /// </summary>
-    Prewitt = 0,
+    PREWITT = 0,
 
     /// <summary>
     /// Sobel operator.
     /// </summary>
-    Sobel = 1,
+    SOBEL = 1,
 
     /// <summary>
     /// Scharr operator.
     /// </summary>
-    Scharr = 2,
+    SCHARR = 2,
 
     /// <summary>
     /// Line Segment Detector operator.
